backend/requirements.txt
Flask
Werkzeug
openai
requests
azure-identity
python-dotenv
Pillow
numpy
pandas